+++ ProjectWebApp/Controllers/MENU_SUBController.cs
@@ -17,37 +17,68 @@ namespace LexAbogadosWeb.Controllers
         // GET: MENU_SUB
         public ActionResult Index()
         {
-            var mENU_SUB = db.MENU_SUB.Include(m => m.MENU).Include(m => m.TIPO_ROL);
-            return View(mENU_SUB.ToList());
+            List<MENU_SUB> mENU_SUB = new List<MENU_SUB>();
+            try
+            {
+                mENU_SUB = db.MENU_SUB.Include(m => m.MENU).Include(m => m.TIPO_ROL).ToList();
+                return View(mENU_SUB);
+            }
+            catch (Exception)
+            {
+
+                return View(mENU_SUB);
+            }
+
         }
 
         // GET: MENU_SUB/Details/5
         public ActionResult Details(long? id)
         {
-            if (id == null)
+            MENU_SUB mENU_SUB = new MENU_SUB();
+            try
             {
-                return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+                if (id == null)
+                {
+                    return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+                }
+                mENU_SUB = db.MENU_SUB.Find(id);
+                if (mENU_SUB == null)
+                {
+                    return HttpNotFound();
+                }
+                return View(mENU_SUB);
             }
-            MENU_SUB mENU_SUB = db.MENU_SUB.Find(id);
-            if (mENU_SUB == null)
+            catch (Exception)
             {
-                return HttpNotFound();
+
+                return View(mENU_SUB);
             }
-            return View(mENU_SUB);
+
         }
 
         // GET: MENU_SUB/Create
         public ActionResult Create()
         {
-            using (ODAO Menu = new ODAO())
+            try
             {
-                
-                ViewBag.SUB = new SelectList(Menu.MENU.Include("TIPO_ROL").ToList(), "MENU_ID", "TEXTO", "TIPO_ROL.NOMBRE_ROL", 1);
+                using (ODAO Menu = new ODAO())
+                {
+
+                    ViewBag.SUB = new SelectList(Menu.MENU.Include("TIPO_ROL").ToList(), "MENU_ID", "TEXTO", "TIPO_ROL.NOMBRE_ROL", 1);
+                }
+
+                ViewBag.MENU_ID_P = new SelectList(db.MENU, "MENU_ID", "TEXTO", "");
+                ViewBag.ID_ROL = new SelectList(db.TIPO_ROL, "ID_ROL", "NOMBRE_ROL");
+                return View();
             }
-            
-            ViewBag.MENU_ID_P = new SelectList(db.MENU, "MENU_ID", "TEXTO","");
-            ViewBag.ID_ROL = new SelectList(db.TIPO_ROL, "ID_ROL", "NOMBRE_ROL");
-            return View();
+            catch (Exception)
+            {
+                ViewBag.SUB = new SelectList(new List<string>());
+                ViewBag.MENU_ID_P = new SelectList( new List<string>());
+                ViewBag.ID_ROL = new SelectList(new List<string>());
+                return View();
+            }
+
         }
 
         // POST: MENU_SUB/Create
@@ -57,33 +88,52 @@ namespace LexAbogadosWeb.Controllers
         [ValidateAntiForgeryToken]
         public ActionResult Create([Bind(Include = "MENU_ID,TEXTO,REFERENCIA,ID_ROL,MENU_ID_P,CRUD,CONTROLLER,ACTION")] MENU_SUB mENU_SUB)
         {
-            if (ModelState.IsValid)
+            try
             {
-                db.MENU_SUB.Add(mENU_SUB);
-                db.SaveChanges();
-                return RedirectToAction("Index");
+                if (ModelState.IsValid)
+                {
+                    db.MENU_SUB.Add(mENU_SUB);
+                    db.SaveChanges();
+                    return RedirectToAction("Index");
+                }
+
+                ViewBag.MENU_ID_P = new SelectList(db.MENU, "MENU_ID", "TEXTO", mENU_SUB.MENU_ID_P);
+                ViewBag.ID_ROL = new SelectList(db.TIPO_ROL, "ID_ROL", "NOMBRE_ROL", mENU_SUB.ID_ROL);
+                return View(mENU_SUB);
+            }
+            catch (Exception)
+            {
+
+                return View(mENU_SUB);
             }
 
-            ViewBag.MENU_ID_P = new SelectList(db.MENU, "MENU_ID", "TEXTO", mENU_SUB.MENU_ID_P);
-            ViewBag.ID_ROL = new SelectList(db.TIPO_ROL, "ID_ROL", "NOMBRE_ROL", mENU_SUB.ID_ROL);
-            return View(mENU_SUB);
         }
 
         // GET: MENU_SUB/Edit/5
         public ActionResult Edit(long? id)
         {
-            if (id == null)
+            MENU_SUB mENU_SUB = new MENU_SUB();
+            try
             {
-                return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+                if (id == null)
+                {
+                    return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+                }
+                mENU_SUB = db.MENU_SUB.Find(id);
+                if (mENU_SUB == null)
+                {
+                    return HttpNotFound();
+                }
+                ViewBag.MENU_ID_P = new SelectList(db.MENU, "MENU_ID", "TEXTO", mENU_SUB.MENU_ID_P);
+                ViewBag.ID_ROL = new SelectList(db.TIPO_ROL, "ID_ROL", "NOMBRE_ROL", mENU_SUB.ID_ROL);
+                return View(mENU_SUB);
             }
-            MENU_SUB mENU_SUB = db.MENU_SUB.Find(id);
-            if (mENU_SUB == null)
+            catch (Exception)
             {
-                return HttpNotFound();
+                ViewBag.MENU_ID_P = new SelectList(new List<string>());
+                ViewBag.ID_ROL = new SelectList(new List<string>());
+                return View(mENU_SUB);
             }
-            ViewBag.MENU_ID_P = new SelectList(db.MENU, "MENU_ID", "TEXTO", mENU_SUB.MENU_ID_P);
-            ViewBag.ID_ROL = new SelectList(db.TIPO_ROL, "ID_ROL", "NOMBRE_ROL", mENU_SUB.ID_ROL);
-            return View(mENU_SUB);
         }
 
         // POST: MENU_SUB/Edit/5
@@ -93,30 +143,50 @@ namespace LexAbogadosWeb.Controllers
         [ValidateAntiForgeryToken]
         public ActionResult Edit([Bind(Include = "MENU_ID,TEXTO,REFERENCIA,ID_ROL,MENU_ID_P,CRUD,CONTROLLER,ACTION")] MENU_SUB mENU_SUB)
         {
-            if (ModelState.IsValid)
+            try
             {
-                db.Entry(mENU_SUB).State = EntityState.Modified;
-                db.SaveChanges();
-                return RedirectToAction("Index");
+                if (ModelState.IsValid)
+                {
+                    db.Entry(mENU_SUB).State = EntityState.Modified;
+                    db.SaveChanges();
+                    return RedirectToAction("Index");
+                }
+                ViewBag.MENU_ID_P = new SelectList(db.MENU, "MENU_ID", "TEXTO", mENU_SUB.MENU_ID_P);
+                ViewBag.ID_ROL = new SelectList(db.TIPO_ROL, "ID_ROL", "NOMBRE_ROL", mENU_SUB.ID_ROL);
+                return View(mENU_SUB);
             }
-            ViewBag.MENU_ID_P = new SelectList(db.MENU, "MENU_ID", "TEXTO", mENU_SUB.MENU_ID_P);
-            ViewBag.ID_ROL = new SelectList(db.TIPO_ROL, "ID_ROL", "NOMBRE_ROL", mENU_SUB.ID_ROL);
-            return View(mENU_SUB);
+            catch (Exception)
+            {
+                ViewBag.MENU_ID_P = new SelectList(new List<string>());
+                ViewBag.ID_ROL = new SelectList(new List<string>());
+                return View(mENU_SUB);
+            }
+
         }
 
         // GET: MENU_SUB/Delete/5
         public ActionResult Delete(long? id)
         {
-            if (id == null)
+            MENU_SUB mENU_SUB = new MENU_SUB();
+            try
             {
-                return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+                if (id == null)
+                {
+                    return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+                }
+                mENU_SUB = db.MENU_SUB.Find(id);
+                if (mENU_SUB == null)
+                {
+                    return HttpNotFound();
+                }
+                return View(mENU_SUB);
             }
-            MENU_SUB mENU_SUB = db.MENU_SUB.Find(id);
-            if (mENU_SUB == null)
+            catch (Exception)
             {
-                return HttpNotFound();
+
+                return View(mENU_SUB);
             }
-            return View(mENU_SUB);
+
         }
 
         // POST: MENU_SUB/Delete/5
@@ -124,10 +194,22 @@ namespace LexAbogadosWeb.Controllers
         [ValidateAntiForgeryToken]
         public ActionResult DeleteConfirmed(long id)
         {
-            MENU_SUB mENU_SUB = db.MENU_SUB.Find(id);
-            db.MENU_SUB.Remove(mENU_SUB);
-            db.SaveChanges();
-            return RedirectToAction("Index");
+            try
+            {
+                if (id > 0)
+                {
+                    MENU_SUB mENU_SUB = db.MENU_SUB.Find(id);
+                    db.MENU_SUB.Remove(mENU_SUB);
+                    db.SaveChanges();
+                }
+                return RedirectToAction("Index");
+            }
+            catch (Exception)
+            {
+
+                return RedirectToAction("Index");
+            }
+
         }
 
         protected override void Dispose(bool disposing)
